requirements.txt
requests
python-dotenv
fastapi
uvicorn
httpx
pytest